+++ src/HemiFrame/Lib/Cache/Memcached.php
@@ -4,13 +4,9 @@ namespace HemiFrame\Lib\Cache;
 
 class Memcached implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterface
 {
-    private $keyPrefix = '';
-    private $defaultTtl = 120;
-
-    /**
-     * @var \Memcached
-     */
-    private $memcached;
+    private string $keyPrefix = '';
+    private int $defaultTtl = 120;
+    private \Memcached $memcached;
 
     public function getKeyPrefix(): string
     {
@@ -37,27 +33,24 @@ class Memcached implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheIn
     }
 
     /**
-     * @param string $key
-     * @param int    $time
-     *
      * @throws InvalidArgumentException
      */
-    public function set($key, $value, $time = null): bool
+    public function set(string $key, mixed $value, int|\DateInterval|null $ttl = null): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
         }
-        if (null === $time) {
-            $time = $this->defaultTtl;
+        if ($ttl instanceof \DateInterval) {
+            $ttl = (int) (new \DateTime())->add($ttl)->getTimestamp() - time();
+        }
+        if (null === $ttl) {
+            $ttl = $this->defaultTtl;
         }
 
-        return $this->memcached->set($this->keyPrefix.$key, $value, $time);
+        return $this->memcached->set($this->keyPrefix.$key, $value, $ttl);
     }
 
-    /**
-     * @param string $key
-     */
-    public function get($key, $default = null)
+    public function get(string $key, mixed $default = null): mixed
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Key is empty');
@@ -72,11 +65,9 @@ class Memcached implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheIn
     }
 
     /**
-     * @param string $key
-     *
      * @throws InvalidArgumentException
      */
-    public function delete($key): bool
+    public function delete(string $key): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
@@ -91,11 +82,9 @@ class Memcached implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheIn
     }
 
     /**
-     * @param string $key
-     *
      * @throws InvalidArgumentException
      */
-    public function has($key): bool
+    public function has(string $key): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
@@ -117,10 +106,7 @@ class Memcached implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheIn
         return $this->has($key);
     }
 
-    /**
-     * @param array|mixed $keys
-     */
-    public function getMultiple($keys, $default = null): array
+    public function getMultiple(iterable $keys, mixed $default = null): array
     {
         if (!is_array($keys)) {
             throw new InvalidArgumentException('Keys must be array');
@@ -134,7 +120,7 @@ class Memcached implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheIn
         return $data;
     }
 
-    public function setMultiple($values, $ttl = null): bool
+    public function setMultiple(iterable $values, int|\DateInterval|null $ttl = null): bool
     {
         if (!is_array($values)) {
             throw new InvalidArgumentException('Values must be array');
@@ -150,7 +136,7 @@ class Memcached implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheIn
         return $result;
     }
 
-    public function deleteMultiple($keys): bool
+    public function deleteMultiple(iterable $keys): bool
     {
         if (!is_array($keys)) {
             throw new InvalidArgumentException('Keys must be array');