+++ src/HemiFrame/Lib/Cache/Memory.php
@@ -7,46 +7,43 @@ namespace HemiFrame\Lib\Cache;
  */
 class Memory implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterface
 {
-    private static $data = [];
-    private $keyPrefix = '';
-    private $defaultTtl = 120;
+    private static array $data = [];
+    private string $keyPrefix = '';
+    private int $defaultTtl = 120;
 
     public function getKeyPrefix(): string
     {
         return $this->keyPrefix;
     }
 
-    public function setKeyPrefix(string $keyPrefix)
+    public function setKeyPrefix(string $keyPrefix): void
     {
         $this->keyPrefix = $keyPrefix;
     }
 
     /**
-     * @param string $key
-     * @param int    $time
-     *
      * @throws InvalidArgumentException
      */
-    public function set($key, $value, $time = null): bool
+    public function set(string $key, mixed $value, int|\DateInterval|null $ttl = null): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
         }
-        if (null === $time) {
-            $time = $this->defaultTtl;
+        if ($ttl instanceof \DateInterval) {
+            $ttl = (int) (new \DateTime())->add($ttl)->getTimestamp() - time();
+        }
+        if (null === $ttl) {
+            $ttl = $this->defaultTtl;
         }
         self::$data[$this->keyPrefix.$key] = [
-            'expiryTime' => time() + $time,
+            'expiryTime' => time() + $ttl,
             'value' => $value,
         ];
 
         return true;
     }
 
-    /**
-     * @param string $key
-     */
-    public function get($key, $default = null)
+    public function get(string $key, mixed $default = null): mixed
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Key is empty');
@@ -61,11 +58,9 @@ class Memory implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInter
     }
 
     /**
-     * @param string $key
-     *
      * @throws InvalidArgumentException
      */
-    public function delete($key): bool
+    public function delete(string $key): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
@@ -83,11 +78,9 @@ class Memory implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInter
     }
 
     /**
-     * @param string $key
-     *
      * @throws InvalidArgumentException
      */
-    public function has($key): bool
+    public function has(string $key): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
@@ -109,10 +102,7 @@ class Memory implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInter
         return $this->has($key);
     }
 
-    /**
-     * @param array|mixed $keys
-     */
-    public function getMultiple($keys, $default = null): array
+    public function getMultiple(iterable $keys, mixed $default = null): array
     {
         if (!is_array($keys)) {
             throw new InvalidArgumentException('Keys must be array');
@@ -126,7 +116,7 @@ class Memory implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInter
         return $data;
     }
 
-    public function setMultiple($values, $ttl = null): bool
+    public function setMultiple(iterable $values, int|\DateInterval|null $ttl = null): bool
     {
         if (!is_array($values)) {
             throw new InvalidArgumentException('Values must be array');
@@ -142,7 +132,7 @@ class Memory implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInter
         return $result;
     }
 
-    public function deleteMultiple($keys): bool
+    public function deleteMultiple(iterable $keys): bool
     {
         if (!is_array($keys)) {
             throw new InvalidArgumentException('Keys must be array');