+++ src/HemiFrame/Lib/Cache/Redis.php
@@ -5,8 +5,8 @@ namespace HemiFrame\Lib\Cache;
 class Redis implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterface
 {
     private \Redis $redis;
-    private $keyPrefix = '';
-    private $defaultTtl = 120;
+    private string $keyPrefix = '';
+    private int $defaultTtl = 120;
 
     public function __construct(\Redis $redis)
     {
@@ -26,27 +26,24 @@ class Redis implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterf
     }
 
     /**
-     * @param string $key
-     * @param int    $time
-     *
      * @throws InvalidArgumentException
      */
-    public function set($key, $value, $time = null): bool
+    public function set(string $key, mixed $value, int|\DateInterval|null $ttl = null): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
         }
-        if (null === $time) {
-            $time = $this->defaultTtl;
+        if ($ttl instanceof \DateInterval) {
+            $ttl = (int) (new \DateTime())->add($ttl)->getTimestamp() - time();
+        }
+        if (null === $ttl) {
+            $ttl = $this->defaultTtl;
         }
 
-        return $this->redis->set($this->keyPrefix.$key, serialize($value), ['ex' => $time]);
+        return $this->redis->set($this->keyPrefix.$key, serialize($value), ['ex' => $ttl]);
     }
 
-    /**
-     * @param string $key
-     */
-    public function get($key, $default = null)
+    public function get(string $key, mixed $default = null): mixed
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Key is empty');
@@ -60,11 +57,9 @@ class Redis implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterf
     }
 
     /**
-     * @param string $key
-     *
      * @throws InvalidArgumentException
      */
-    public function delete($key): bool
+    public function delete(string $key): bool
     {
         if (empty($key)) {
             throw new InvalidArgumentException('Enter key');
@@ -111,10 +106,7 @@ class Redis implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterf
         return $this->has($key);
     }
 
-    /**
-     * @param array|mixed $keys
-     */
-    public function getMultiple($keys, $default = null): array
+    public function getMultiple(iterable $keys, mixed $default = null): array
     {
         if (!is_array($keys)) {
             throw new InvalidArgumentException('Keys must be array');
@@ -131,7 +123,7 @@ class Redis implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterf
         return $data;
     }
 
-    public function setMultiple($values, $ttl = null): bool
+    public function setMultiple(iterable $values, int|\DateInterval|null $ttl = null): bool
     {
         if (!is_array($values)) {
             throw new InvalidArgumentException('Values must be array');
@@ -147,7 +139,7 @@ class Redis implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterf
         return $result;
     }
 
-    public function deleteMultiple($keys): bool
+    public function deleteMultiple(iterable $keys): bool
     {
         if (!is_array($keys)) {
             throw new InvalidArgumentException('Keys must be array');
@@ -159,7 +151,7 @@ class Redis implements \HemiFrame\Interfaces\Cache, \Psr\SimpleCache\CacheInterf
         return true;
     }
 
-    public function getRedis()
+    public function getRedis(): \Redis
     {
         return $this->redis;
     }